requirements.txt
dnspython==2.4.2
Flask==3.0.0
Flask-RESTful==0.3.10
pymongo==4.5.0
python-dotenv==1.0.0
qrcode